% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{add_r_squared}
\alias{add_r_squared}
\title{Add R-squared values to endogenous nodes}
\usage{
add_r_squared(inside_node = TRUE)
}
\arguments{
\item{inside_node}{Logical, whether to display R-squared inside nodes
(appended to label) or as external annotations (default TRUE)}
}
\value{
A modifier that can be added to a ggsemplot plot with \code{+}
}
\description{
Add R-squared values to endogenous nodes
}
