% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_sem_model}
\alias{assert_sem_model}
\title{Validate that an object is a sem_model}
\usage{
assert_sem_model(x, arg_name = "model")
}
\description{
Validate that an object is a sem_model
}
\keyword{internal}
