% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-edges.R
\name{.compute_covariance_segments}
\alias{.compute_covariance_segments}
\title{Compute covariance curve endpoints}
\usage{
.compute_covariance_segments(cov_edges, nodes, node_sizes)
}
\description{
Compute covariance curve endpoints
}
\keyword{internal}
