% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-labels.R
\name{.compute_label_positions}
\alias{.compute_label_positions}
\title{Compute label positions at midpoints of edges}
\usage{
.compute_label_positions(edges, nodes, what, sig_stars)
}
\description{
Compute label positions at midpoints of edges
}
\keyword{internal}
