% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-nodes.R
\name{.rounded_rect_polygon}
\alias{.rounded_rect_polygon}
\title{Generate polygon data for a rounded rectangle}
\usage{
.rounded_rect_polygon(cx, cy, hw, hh, r, n_corner = 15)
}
\description{
Generate polygon data for a rounded rectangle
}
\keyword{internal}
