% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ellipse_boundary}
\alias{ellipse_boundary}
\title{Compute the intersection point of a line from an external point to the
center of an ellipse, on the ellipse boundary.}
\usage{
ellipse_boundary(cx, cy, a, b, px, py)
}
\arguments{
\item{cx, cy}{Center of ellipse}

\item{a, b}{Semi-major and semi-minor axes}

\item{px, py}{External point}
}
\value{
Named numeric vector with x and y
}
\description{
Compute the intersection point of a line from an external point to the
center of an ellipse, on the ellipse boundary.
}
\keyword{internal}
