% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggplot_add.sem_palette_modifier}
\alias{ggplot_add.sem_palette_modifier}
\title{Apply palette modifier to a ggplot (via ggplot_add)}
\usage{
\method{ggplot_add}{sem_palette_modifier}(object, plot, object_name)
}
\description{
Apply palette modifier to a ggplot (via ggplot_add)
}
\keyword{internal}
