% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{node_boundary}
\alias{node_boundary}
\title{Get boundary point of a node given its type}
\usage{
node_boundary(node, px, py, node_sizes)
}
\arguments{
\item{node}{A single row from the nodes data.frame}

\item{px, py}{Target point direction}

\item{node_sizes}{Named list with latent_a, latent_b, obs_hw, obs_hh}
}
\value{
Named numeric vector with x, y
}
\description{
Get boundary point of a node given its type
}
\keyword{internal}
