% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-utils.R
\name{nudge_node}
\alias{nudge_node}
\title{Nudge a node's position}
\usage{
nudge_node(model, node_name, dx = 0, dy = 0)
}
\arguments{
\item{model}{A \link{sem_model} object (must have layout computed)}

\item{node_name}{Character name of the node}

\item{dx}{Horizontal shift (default 0)}

\item{dy}{Vertical shift (default 0)}
}
\value{
Modified sem_model
}
\description{
Shift a node by dx/dy relative to its current position.
}
