% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rect_boundary}
\alias{rect_boundary}
\title{Compute intersection of a line with a rectangle boundary}
\usage{
rect_boundary(cx, cy, hw, hh, px, py)
}
\arguments{
\item{cx, cy}{Center of rectangle}

\item{hw, hh}{Half-width and half-height}

\item{px, py}{External point}
}
\value{
Named numeric vector with x and y
}
\description{
Compute intersection of a line with a rectangle boundary
}
\keyword{internal}
