% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-loops.R
\name{render_loops}
\alias{render_loops}
\title{Build ggplot2 layers for residual variance loops}
\usage{
render_loops(edges, nodes, sem_theme)
}
\arguments{
\item{edges}{Data.frame of edges}

\item{nodes}{Data.frame of nodes with layout coordinates}

\item{sem_theme}{A sem_theme list}
}
\value{
A list of ggplot2 layers
}
\description{
Draws small arcs at nodes to indicate residual (error) variance.
}
\keyword{internal}
