% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{style_node}
\alias{style_node}
\title{Customize the style of a specific node}
\usage{
style_node(node_name, fill = NULL, color = NULL, label = NULL)
}
\arguments{
\item{node_name}{Name of the node to style}

\item{fill}{Fill color (NULL to keep current)}

\item{color}{Border color (NULL to keep current)}

\item{label}{New display label (NULL to keep current)}
}
\value{
A modifier that can be added to a ggsemplot plot with \code{+}
}
\description{
Customize the style of a specific node
}
