% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-utils.R
\name{swap_nodes}
\alias{swap_nodes}
\title{Swap positions of two nodes}
\usage{
swap_nodes(model, node1, node2)
}
\arguments{
\item{model}{A \link{sem_model} object (must have layout computed)}

\item{node1}{Name of first node}

\item{node2}{Name of second node}
}
\value{
Modified sem_model
}
\description{
Swap positions of two nodes
}
